fix(rest): AgentUsersItem.PasswordLastChange must be float64, not int64
Refs rimor-plugin-scanner-wazuh#39.

The Wazuh syscollector agents-users API returns password_last_change as a
fractional Unix timestamp (e.g. 1787316174.6758862), so decoding into an int64
field failed with "json: cannot unmarshal number ... into Go struct field
SyscollectorAgentUsersResponse.data.affected_items.N.user.password_last_change of
type int64", aborting the whole agents-users query. Widened the field to float64
(which also accepts integer timestamps). Decode test added.

Note: consumers that treat this as an int64 (the wazuh scanner's
codex.TimestampFromUnix call) must adapt when they bump this dependency.

+++ rest/types.go
@@ -8308,23 +8308,23 @@ type AgentUsersScan struct {
 }
 
 type AgentUsersItem struct {
-	AuthFailedCount                 int    `json:"auth_failed_count"`
-	GroupID                         int    `json:"group_id"`
-	GroupIDSigned                   int    `json:"group_id_signed"`
-	Groups                          string `json:"groups"`
-	Home                            string `json:"home"`
-	ID                              int    `json:"id"`
-	IsHidden                        int    `json:"is_hidden"`
-	IsRemote                        int    `json:"is_remote"`
-	Name                            string `json:"name"`
-	PasswordLastChange              int64  `json:"password_last_change"`
-	PasswordMaxDaysBetweenChanges   int    `json:"password_max_days_between_changes"`
-	PasswordMinDaysBetweenChanges   int    `json:"password_min_days_between_changes"`
-	PasswordStatus                  string `json:"password_status"`
-	PasswordWarningDaysBeforeExpire int    `json:"password_warning_days_before_expiration"`
-	Roles                           string `json:"roles"`
-	Shell                           string `json:"shell"`
-	Type                            string `json:"type"`
-	UIDSigned                       int    `json:"uid_signed"`
-	UUID                            string `json:"uuid"`
+	AuthFailedCount                 int     `json:"auth_failed_count"`
+	GroupID                         int     `json:"group_id"`
+	GroupIDSigned                   int     `json:"group_id_signed"`
+	Groups                          string  `json:"groups"`
+	Home                            string  `json:"home"`
+	ID                              int     `json:"id"`
+	IsHidden                        int     `json:"is_hidden"`
+	IsRemote                        int     `json:"is_remote"`
+	Name                            string  `json:"name"`
+	PasswordLastChange              float64 `json:"password_last_change"`
+	PasswordMaxDaysBetweenChanges   int     `json:"password_max_days_between_changes"`
+	PasswordMinDaysBetweenChanges   int     `json:"password_min_days_between_changes"`
+	PasswordStatus                  string  `json:"password_status"`
+	PasswordWarningDaysBeforeExpire int     `json:"password_warning_days_before_expiration"`
+	Roles                           string  `json:"roles"`
+	Shell                           string  `json:"shell"`
+	Type                            string  `json:"type"`
+	UIDSigned                       int     `json:"uid_signed"`
+	UUID                            string  `json:"uuid"`
 }
